refactor: renamed themes.ts and routes.ts files
