+++ src/constants/appTheme.ts
@@ -1,4 +1,4 @@
-export enum AppThemes {
+export enum AppTheme {
   LIGHT = 'light',
   DARK = 'dark',
 }
