+++ src/constants/index.ts
@@ -1,3 +1,5 @@
+export { AppRoutes } from './appRoutes';
+export { AppTheme } from './appTheme';
 export { BANKS_DATA } from './banksData';
 export { CONTACTS_INFO } from './contactsInfo';
 export { DEFAULT_CHART_DATA } from './defaultChartData';
@@ -5,9 +7,7 @@ export { DEFAULT_CURRENCIES } from './defaultCurrencies';
 export { FOOTER_LINK_GROUPS } from './footerLinkGroups';
 export { ICONS } from './icons';
 export { PAGE_LIST } from './pageList';
-export { AppRoutes } from './routes';
 export { SOCIAL_LIST } from './socialList';
-export { AppThemes } from './themes';
 
 export const CURRENCY_MODAL_QUERY_KEY = 'currency';
